clean install.sh

+++ install.sh
@@ -2,7 +2,9 @@
 if ! pip show pyperclip &> /dev/null; then
     pip install pyperclip
 fi
-sudo apt-get install xclip #for linux pyperclip
+if ! dpkg -l xclip &> /dev/null; then
+    sudo apt-get install xclip #for linux pyperclip
+fi
 
 if ! pip show python-docx &> /dev/null; then
     pip install python-docx
@@ -11,4 +13,8 @@ fi
 if ! pip show Pillow &> /dev/null; then
     pip install Pillow
 fi
-pip install pyqt5
+
+if ! pip show pyqt5 &> /dev/null; then
+    pip install pyqt5 # -i https://pypi.tuna.tsinghua.edu.cn/simple/
+    # download fast url
+fi
